+++ nextstep-di/src/main/java/nextstep/di/factory/BeanFactoryUtils.java
@@ -2,28 +2,61 @@ package nextstep.di.factory;
 
 import com.google.common.collect.Sets;
 import nextstep.annotation.Inject;
+import nextstep.exception.NotFoundConstructorException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.lang.reflect.Constructor;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Set;
-
-import static org.reflections.ReflectionUtils.getAllConstructors;
-import static org.reflections.ReflectionUtils.withAnnotation;
+import java.util.stream.Collectors;
 
 public class BeanFactoryUtils {
+    private static final Logger log = LoggerFactory.getLogger(BeanFactoryUtils.class);
+
+    private static final int ONE_CONSTRUCTOR = 1;
+    private static final int FIRST_CONSTRUCTOR = 0;
+
     /**
-     * 인자로 전달하는 클래스의 생성자 중 @Inject 애노테이션이 설정되어 있는 생성자를 반환
+     * 생성자가 하나이면 그 생성자 반환, 여러개이면 그중 @Inject가 붙은 생성자가 하나이면 그 생성자 반환
+     * 이외의 경우에는 예외 발생
      *
      * @param clazz
      * @return
-     * @Inject 애노테이션이 설정되어 있는 생성자는 클래스당 하나로 가정한다.
      */
-    @SuppressWarnings({"rawtypes", "unchecked"})
-    public static Constructor<?> getInjectedConstructor(Class<?> clazz) {
-        Set<Constructor> injectedConstructors = getAllConstructors(clazz, withAnnotation(Inject.class));
-        if (injectedConstructors.isEmpty()) {
-            return null;
+    public static Constructor findConstructor(Class<?> clazz) {
+        log.debug("Find Constructor : {}", clazz.getName());
+
+        checkInterface(clazz);
+
+        List<Constructor> constructors = Arrays.asList(clazz.getConstructors());
+
+        if (hasOneConstructor(constructors)) {
+            return constructors.get(FIRST_CONSTRUCTOR);
         }
-        return injectedConstructors.iterator().next();
+        return findInjectedConstructor(constructors);
+    }
+
+    private static void checkInterface(Class<?> clazz) {
+        if (clazz.isInterface()) {
+            throw new NotFoundConstructorException("올바른 생성자를 찾을 수 없습니다.");
+        }
+    }
+
+    private static Constructor findInjectedConstructor(List<Constructor> constructors) {
+        List<Constructor> injectedConstructors = constructors.stream()
+                .filter(constructor -> constructor.isAnnotationPresent(Inject.class))
+                .collect(Collectors.toList());
+
+        if (hasOneConstructor(injectedConstructors)) {
+            return injectedConstructors.get(FIRST_CONSTRUCTOR);
+        }
+        throw new NotFoundConstructorException("올바른 생성자를 찾을 수 없습니다.");
+    }
+
+    private static boolean hasOneConstructor(List<Constructor> constructors) {
+        return constructors.size() == ONE_CONSTRUCTOR;
     }
 
     /**
