+++ nextstep-di/src/test/java/nextstep/di/factory/utilsexample/MultipleController.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.exception.multiple;
+package nextstep.di.factory.utilsexample;
 
 import nextstep.stereotype.Controller;
