+++ nextstep-di/src/test/java/nextstep/di/factory/utilsexample/NotExistConstroller.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.exception.notexist;
+package nextstep.di.factory.utilsexample;
 
 import nextstep.stereotype.Controller;
