+++ nextstep-di/src/test/java/nextstep/di/factory/utilsexample/OneInjectController.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.exception.oneinject;
+package nextstep.di.factory.utilsexample;
 
 import nextstep.annotation.Inject;
 import nextstep.stereotype.Controller;
