+++ nextstep-di/src/test/java/nextstep/di/factory/utilsexample/UtilsInterface.java
@@ -1,0 +1,4 @@
+package nextstep.di.factory.utilsexample;
+
+public interface UtilsInterface {
+}
